+++ app/src/main/java/com/imagedge/camera/feature/edit/video_to_motion/VideoToLivePhotoScreen.kt
@@ -1,4 +1,4 @@
-package com.imagedge.camera.feature.edit
+package com.imagedge.camera.feature.edit.video_to_motion
 
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.PickVisualMediaRequest